+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to the "Code Prism" extension will be documented in this fil
 bug fix
 
 - fixed the bug that don't delete comments when the prism file is deleted
+- fixed the bug that don't append new issue when prism tree view is list mode
 - fixed the bug that don't change the category when prism tree view is list mode
 
 ### 1.3.1 - November 6, 2024